refactor(api): extract testimonial route handlers and types to _lib
Extracts business logic, validation schemas, rate limiting, and database operations from the Hono router file [[path]].ts into a separate _lib.ts module. This keeps the main route declarations clean, isolates Cloudflare bindings and domain types, and simplifies testing.
